create plugin for modals
