+++ resources/js/plugins/Modals.js
@@ -1,0 +1,19 @@
+const Modals = {
+    install(Vue, options) {
+        this.EventBus = new Vue()
+
+        Vue.prototype.$modal = {
+            show(modal, data) {
+                Modals.EventBus.$emit('show::modal::' + modal, modal, data);
+            },
+
+            showRecentInventoryMovementsModal(inventory_id) {
+                this.show('recent-inventory-movements-modal', {
+                    'inventory_id': inventory_id
+                });
+            },
+        }
+    }
+}
+
+export default Modals
